:ok_hand: Refactor the services directory
